+++ packages/at_persistence_spec/lib/at_persistence_spec.dart
@@ -10,4 +10,5 @@ export 'package:at_persistence_spec/src/keystore/log_keystore.dart';
 export 'package:at_persistence_spec/src/keystore/key_pattern.dart';
 export 'package:at_persistence_spec/src/keystore/keystore_change.dart';
 export 'package:at_persistence_spec/src/keystore/keystore_txn.dart';
+export 'package:at_persistence_spec/src/keystore/order_by_key.dart';
 export 'package:at_persistence_spec/src/keystore/secondary_keystore.dart';